+++ hms/src/main/java/com/application/hms/model/MedicalRecord.java
@@ -9,7 +9,7 @@ public class MedicalRecord {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long recordId;
+    private int recordId;
 
     @ManyToOne
     @JoinColumn(name = "patient_id")
